functions/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-api-core
python-multipart